+++ x/observer/client/cli/query_get_tss_address.go
@@ -11,9 +11,9 @@ import (
 
 func CmdGetTssAddress() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "get-tss-address",
+		Use:   "get-tss-address [bitcoinChainId]]",
 		Short: "Query current tss address",
-		Args:  cobra.NoArgs,
+		Args:  cobra.MaximumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
 
 			clientCtx, err := client.GetClientTxContext(cmd)
@@ -21,8 +21,14 @@ func CmdGetTssAddress() *cobra.Command {
 				return err
 			}
 			queryClient := types.NewQueryClient(clientCtx)
-
 			params := &types.QueryGetTssAddressRequest{}
+			if len(args) == 1 {
+				bitcoinChainID, err := strconv.ParseInt(args[0], 10, 64)
+				if err != nil {
+					return err
+				}
+				params.BitcoinChainId = bitcoinChainID
+			}
 
 			res, err := queryClient.GetTssAddress(cmd.Context(), params)
 			if err != nil {
@@ -40,9 +46,9 @@ func CmdGetTssAddress() *cobra.Command {
 
 func CmdGetTssAddressByFinalizedZetaHeight() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "get-historical-tss-address [finalizedZetaHeight]",
+		Use:   "get-historical-tss-address [finalizedZetaHeight] [bitcoinChainId]",
 		Short: "Query tss address by finalized zeta height (for historical tss addresses)",
-		Args:  cobra.ExactArgs(1),
+		Args:  cobra.MaximumNArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
 
 			clientCtx, err := client.GetClientTxContext(cmd)
@@ -57,6 +63,13 @@ func CmdGetTssAddressByFinalizedZetaHeight() *cobra.Command {
 			}
 			params := &types.QueryGetTssAddressByFinalizedHeightRequest{
 				FinalizedZetaHeight: finalizedZetaHeight,
+			}
+			if len(args) == 2 {
+				bitcoinChainID, err := strconv.ParseInt(args[1], 10, 64)
+				if err != nil {
+					return err
+				}
+				params.BitcoinChainId = bitcoinChainID
 			}
 
 			res, err := queryClient.GetTssAddressByFinalizedHeight(cmd.Context(), params)